backend_ml/requirements.txt
dotenv
flask
flask-cors
joblib
pandas
xgboost
scikit-learn
flask_sqlalchemy
pymysql
PyJWT